Redid test framework to bring up driver on instance and expose TCP socket with SSH Tunneling
