+++ test/run-e2e-local.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-set -e
-set -x
+set -o nounset
+set -o errexit
 
 readonly PKGDIR=sigs.k8s.io/gcp-compute-persistent-disk-csi-driver
 
-go run "$GOPATH/src/${PKGDIR}/test/remote/run_remote/run_remote.go" --logtostderr --v 2 --project "${PROJECT}" --zone "${ZONE}" --ssh-env gce --delete-instances=false --cleanup=true --results-dir=my_test  --service-account="${IAM_NAME}"
+go test --v=true "${PKGDIR}/test/e2e" --logtostderr --project ${PROJECT} --service-account ${IAM_NAME}
